little beautiful change from long to Integer in Users

+++ day01/src/ex00/User.java
@@ -7,7 +7,7 @@ public class User {
 
     User() {}
 
-    public long getId() {
+    public Integer getId() {
         return id;
     }
 
@@ -15,7 +15,7 @@ public class User {
         return name;
     }
 
-    public long getBalance() {
+    public Integer getBalance() {
         return balance;
     }
